Extract git tool definitions
