+++ docs/DECISIONS.md
@@ -220,3 +220,4 @@
 - SSH Remote project execution should keep command routing, path validation, and command planning in separate seams. `WorkspaceRemoteProjectToolExecutor` owns only tool routing and SSH execution, `WorkspaceRemoteProjectPath` owns remote path normalization/artifact labels, and `WorkspaceRemoteGitToolRequestPlanner` owns Git/GitHub/worktree command construction plus URL/artifact intent. This keeps remote support testable as SSH, QuillCloud relay, and future remote transports expand.
 - Claude design critique on 2026-06-23 identified the next high-impact UI cleanup: split model selection from approval mode, replace the flat Tools popover with a structured activity surface, teach Auto/Review/Read-only as first-class controls in the empty state, and make approvals a persistent inline queue banner instead of a modal-feeling interruption. Treat those as UX roadmap inputs, not remote-executor scope.
 - Tool-step execution lives in `AgentToolStepRunner.swift`, not in `Agent.swift`. `AgentRunner.send` owns the high-level model/tool loop and repeated-call fallback; the extracted runner owns safety review, tool lifecycle event emission, unavailable-tool results, apply-patch diff follow-up execution, and tool feedback serialization.
+- Git tool schema declarations live in `GitToolDefinitions.swift`, not in `GitToolExecutor.swift`. The executor owns process execution and safety validation; the definitions catalog owns local git, GitHub PR, and worktree tool names, descriptions, JSON schemas, host, and risk metadata.
